+++ app.py
@@ -7,9 +7,10 @@ from entity.model import User, Image
 from entity.model import db
 
 from api.email_api import bp as email_module
+from api.s3_api import bp as s3_module
+
 
 from crypt import methods
-#from urllib import request
 from datetime import datetime as dt
 
 from werkzeug.utils import secure_filename
@@ -22,6 +23,8 @@ app.config.update(DEBUG=True)
 
 app = create_app()
 app.register_blueprint(email_module)
+app.register_blueprint(s3_module)
+
 
 @app.route('/')
 def welcome():
@@ -87,29 +90,29 @@ def fetch_images():
     return json.dumps(all_image), 200
 
 
-@app.route('/s3-image-upload-test', methods=['POST'])
-def s3upload_test():
-    # html에서 가져온 이미지 
-    file = request.files['file']
-    file.save(secure_filename(file.filename))
+# @app.route('/s3-image-upload-test', methods=['POST'])
+# def s3upload_test():
+#     # html에서 가져온 이미지 
+#     file = request.files['file']
+#     file.save(secure_filename(file.filename))
 
-    # 파일 이름 지정
-    filename = file.filename.split('.')[0]
-    image_type = file.filename.split('.')[-1]
-    image_name = dt.now().strftime(f"{filename}--%Y-%m-%d-%H-%M-%S.{image_type}")
+#     # 파일 이름 지정
+#     filename = file.filename.split('.')[0]
+#     image_type = file.filename.split('.')[-1]
+#     image_name = dt.now().strftime(f"{filename}--%Y-%m-%d-%H-%M-%S.{image_type}")
     
-    # s3버킷에 업로드
-    s3_put_object(s3, 'ladder-s3-bucket', file, image_name)
+#     # s3버킷에 업로드
+#     s3_put_object(s3, 'ladder-s3-bucket', file, image_name)
 
-    # 현재 로그인 사용자 정보
-    user_id = 1
+#     # 현재 로그인 사용자 정보
+#     user_id = 1
 
-    # postgres image table에 업로드
-    image_url = "s3://ladder-s3-bucket/images/"+image_name
-    image_url = image_url.replace(" ","/")
-    database.add_instance(Image, user_id = user_id, image_url = image_url)
+#     # postgres image table에 업로드
+#     image_url = "s3://ladder-s3-bucket/images/"+image_name
+#     image_url = image_url.replace(" ","/")
+#     database.add_instance(Image, user_id = user_id, image_url = image_url)
 
-    return "성공적으로 사진이 S3에 저장되었습니다."
+#     return "성공적으로 사진이 S3에 저장되었습니다."
